still changing ints by size_t
git-svn-id: file:///home/roberto/Git/svn@6513 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 6d4849a8a9b807dbeefed3d90e7a678a26e49a0f

+++ trunk/xmipp/libraries/data/rwMRC.h
@@ -125,15 +125,15 @@ int readMRC(int img_select, bool isStack=false)
     // Convert VAX floating point types if necessary
     if ( header->amin > header->amax )
         REPORT_ERROR(ERR_UNCLASSIFIED,"readMRC: amin > max: VAX floating point conversion unsupported");
-    int _xDim,_yDim,_zDim;
-    unsigned long int _nDim;
-    _xDim = (int) header->nx;
-    _yDim = (int) header->ny;
-    _zDim = (int) header->nz;
+    size_t _xDim,_yDim,_zDim;
+    size_t _nDim;
+    _xDim = (size_t) header->nx;
+    _yDim = (size_t) header->ny;
+    _zDim = (size_t) header->nz;
     _nDim = 1;
     if(isStack)
     {
-        _nDim = (unsigned long int) _zDim;
+        _nDim = (size_t) _zDim;
         _zDim = 1;
         replaceNsize=_nDim;
         std::stringstream Num;
